Add a styled labeled checkbox as a template
